server/pyproject.toml
[tool.poetry]
name = "relatorio_ev"
version = "0.1.0"
description = ""
authors = ["Gustavo Henrique Trentini <gustavotrentini22@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
